fixed, aligned TeamPage
