+++ src/App.jsx
@@ -4,7 +4,7 @@ import HomeHero from './HomeHero'
 import StatsSection from './components/StatsSection'
 import MissionSection from './components/MissionSection'
 import ProjectsPage from './ProjectsPage'
-import TeamPage from './TeamPage'
+import TeamPage from './TeamPage/TeamPage'
 import RecruitmentPage from './RecruitmentPage'
 import SponsorshipPage from './SponsorshipPage'
 import WhoWeAreSection from './components/WhoWeAreSection'
